+++ backend/graphql/resolvers/Event.js
@@ -1,30 +1,7 @@
 require('dotenv').config();
 const Event = require('../../models/Event');
-
+const ResolverHelpers = require('./ResolverHelpers.js');
 const { ValidationError } = require('apollo-server-express');
-
-const action_str = {
-	created: 'created',
-	edited: 'edited',
-	deleted: 'deleted',
-	liked: 'liked',
-	unliked: 'unliked',
-	joined: 'joined',
-	left: 'left',
-	moved: 'moved',
-	subscribed: 'subscribed',
-	unsubscribed: 'unsubscribed',
-	invited: 'invited'
-};
-const object_str = {
-	message: 'message',
-	msgComment: 'message comment',
-	folder: 'folder',
-	document: 'document',
-	docComment: 'document comment',
-	team: 'team',
-	user: 'user'
-};
 
 const eventResolver = {
 	Query: {
@@ -45,10 +22,15 @@ const eventResolver = {
 
 	Mutation: {
 		addEvent: (_, { input }) => {
-			console.log('hey nedim, ur a cool guy.');
-			new Event({ ...input }).save().then(event => {
-				console.log(input);
-				event.populate('team').execPopulate();
+			console.log('hey nedim, ur an ice cube kind of guy.');
+			new Event(input).save().then(event => {
+				// console.log(event);
+				// let {_id} = event;
+				// console.log(_id);
+				// _id = _id.toString()
+
+				//event.populate('team').execPopulate();
+				return event;
 			});
 		},
 		deleteEvent: (_, { input: { id } }) =>
@@ -63,4 +45,4 @@ const eventResolver = {
 	}
 };
 
-module.exports = { eventResolver, object_str, action_str };
+module.exports = eventResolver;
